Plugins!
This is the scaffolding work to support running updates via plugin. The
initial core plugin, `uupd`, encapsulates the original functionality of
the app but will be deprecated over the coming days/weeks with
individual plugins for `bootc`, `rpm-ostree`, `brew`, `flatpak` and
`distrobox`.

This will allow us to a) capture stdout for each update and provide more
a more granular/detailed progress bar for status, and b) the ability to
add new updaters (`uv`, `pip`, `incus`, etc) without changes to the core
application.